requirements.txt
easygui==0.98.3
numpy==1.21.5
opencv_contrib_python==4.6.0.66
Pillow==9.3.0
